Call mutations in a bam-file

This function evaluate the presence (calls) of individual mutations from a predefined list.

dreams_vc(
  mutations_df,
  bam_file_path,
  reference_path,
  model,
  alpha = 0.05,
  use_turboem = TRUE,
  calculate_confidence_intervals = FALSE,
  batch_size = NULL
)

Arguments

mutations_df

A data.frame() with candidate mutations (SNVs) (chromosome, positions, reference and alternative)

bam_file_path

Path to .BAM-file

reference_path

Path to reference genome e.g. FASTA-file.

model

A dreams model. See train_dreams_model().

alpha

Alpha-level used for testing and confidence intervals. Default is 0.05.

use_turboem

Logical. Should turboEM::turboem() be used for EM algorithm? Default is TRUE.

calculate_confidence_intervals

Logical. Should confidence intervals be calculated? Default is FALSE.

batch_size

Number of positions to process at a time

Value

A data.frame() with information about the individual mutation calls, including:

chr, genomic_pos

The genomic position of the mutation.

ref, alt

The reference and alternative allele.

EM_converged

If the EM algorithm converged.

EM_steps, fpeval, objfeval

Number of steps and function evaluations by the EM algorithm.

tf_est

The estiamted tumor fraction (allele fraction).

tf_min, tf_max

The confidence interval of tf_est.

exp_count

The expected count of the alternative allele under the error (null) model.

count

The count of the alternative allele.

coverage

The coverage used by the model (only referenceredas with and alternative allele).

full_coverage

The total coverage of the position (for reference).

obs_freq

The observed frequency of the alternative allele.

ll_0, ll_A

The value of the log-likelihood function under the null (tf=0) and alternative (tf>0) hypothesis.

Q_val, df, p_val

The chisq test statistic, degrees of freedom and p-value of the statistical test.

mutation_detected

Whether the mutation was detected at the supplied alpha level.

See also